pyproject.toml
[project]
name = "private-form-mutual-aid"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.6",
    "django==5.2",
    "django-axes>=8.3.1",
    "django-csp>=4.0",
    "django-filter>=25.2",
    "django-import-export>=4.4.0",
    "django-otp>=1.7.0",
    "django-simple-history>=3.11.0",
    "django-two-factor-auth>=1.18.1",
    "gunicorn>=25.3.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "pip-audit>=2.10.0",
    "ruff>=0.15.9",
]

[tool.ruff]
target-version = "py313"
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B", "UP"]
ignore = [
    "S101",  # assert statements are fine in tests
    "S105",  # hardcoded passwords — bandit handles this better
]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*" = ["E501"]
"*/settings/*" = ["F403", "F401"]  # wildcard imports in settings files

[tool.bandit]
exclude_dirs = [".venv", "*/migrations/*"]
skips = ["B101"]  # assert statements


dev = [
    "bandit>=1.9.4",
    "pip-audit>=2.10.0",
    "ruff>=0.15.9",
]
